# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LaraBaseX API Documentation",
        "description": "A comprehensive API documentation for LaraBaseX - Laravel 12 + ReactJS Full Stack Boilerplate",
        "contact": {
            "email": "admin@larabasex.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "LaraBaseX API Server"
        }
    ],
    "paths": {
...
        "/api/sliders": {
            "post": {
                "tags": [
                    "Home"
                ],
                "summary": "Get home page sliders",
                "description": "Retrieve all active sliders for the home page",
                "operationId": "getSliders",
                "responses": {
                    "200": {
                        "description": "Sliders retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "Slider ID",
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "description": "Slider title",
                                                "type": "string",
                                                "example": "Welcome to LaraBaseX"
                                            },
                                            "description": {
                                                "description": "Slider description",
                                                "type": "string",
                                                "example": "Best platform for your needs"
                                            },
                                            "image": {
                                                "description": "Slider image path",
                                                "type": "string",
                                                "example": "sliders/slider1.jpg"
                                            },
                                            "link": {
                                                "description": "Optional link URL",
                                                "type": "string",
                                                "example": "https://example.com",
                                                "nullable": true
                                            },
                                            "is_active": {
                                                "description": "Whether slider is active",
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "order": {
                                                "description": "Display order",
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "basicAuth": {
                "type": "http",
                "description": "Basic authentication for public endpoints",
                "scheme": "basic"
            }
        }
    },
...
}